+++ src/circuit.rs
@@ -661,7 +661,7 @@ fn test_input_circuit_with_bls12_381() {
             nsk: jubjub::Fr::random(&mut rng),
         };
 
-        let viewing_key = proof_generation_key.to_viewing_key();
+        let viewing_key = proof_generation_key.to_viewing_key().unwrap();
 
         let payment_address;
 
@@ -728,7 +728,7 @@ fn test_input_circuit_with_bls12_381() {
                 }
             }
 
-            let expected_nf = note.nf(&viewing_key.nk, position);
+            let expected_nf = note.nf(viewing_key.nk(), position);
             let expected_nf = multipack::bytes_to_bits_le(&expected_nf.0);
             let expected_nf = multipack::compute_multipacking(&expected_nf);
             assert_eq!(expected_nf.len(), 2);
@@ -836,7 +836,7 @@ fn test_input_circuit_with_bls12_381_external_test_vectors() {
             nsk: jubjub::Fr::random(&mut rng),
         };
 
-        let viewing_key = proof_generation_key.to_viewing_key();
+        let viewing_key = proof_generation_key.to_viewing_key().unwrap();
 
         let payment_address;
 
@@ -911,7 +911,7 @@ fn test_input_circuit_with_bls12_381_external_test_vectors() {
                 }
             }
 
-            let expected_nf = note.nf(&viewing_key.nk, position);
+            let expected_nf = note.nf(viewing_key.nk(), position);
             let expected_nf = multipack::bytes_to_bits_le(&expected_nf.0);
             let expected_nf = multipack::compute_multipacking(&expected_nf);
             assert_eq!(expected_nf.len(), 2);
@@ -989,7 +989,7 @@ fn test_output_circuit_with_bls12_381() {
 
         let proof_generation_key = ProofGenerationKey { ak, nsk };
 
-        let viewing_key = proof_generation_key.to_viewing_key();
+        let viewing_key = proof_generation_key.to_viewing_key().unwrap();
 
         let payment_address;